04-64 Props forwarding
