+++ 04-essentials-deepdive/src/components/TabButton.jsx
@@ -1,8 +1,8 @@
-export default function TabButton({children, onSelect, isSelected}) {
+export default function TabButton({children, isSelected, ...props}) {
 
     return ( 
         <li>
-            <button className={isSelected ? 'active' : undefined} onClick={onSelect}>{children}</button>
+            <button className={isSelected ? 'active' : undefined} {...props}>{children}</button>
         </li>
     );
 }
